feat(grow→polish): migrate atmospheric + path_arcs to POLISH 5e/5f (#1368, PR B)
Second of three migration PRs for issue #1368.

Per the structural-vs-narrative principle from PR #1366:
- Atmospheric annotation is sensory grounding for prose — narrative-prep,
  POLISH territory (spec §Phase 5e, R-5e.1 through R-5e.3).
- Path thematic annotation is per-path emotional identity for FILL/DRESS
  framing — narrative-prep, POLISH territory (spec §Phase 5f, R-5f.1
  through R-5f.3).

## What changed

**POLISH** (`polish/llm_phases.py`):
- New `_phase_5e_atmospheric` decorated with
  `@polish_phase(name="atmospheric_annotation", depends_on=["character_arcs"], priority=3)`.
  Single LLM call generates `atmospheric_detail` for every beat in
  the frozen DAG. Logs `phase5e_partial_coverage` WARNING per R-5e.1
  if any beats end up un-annotated.
- New `_phase_5f_path_thematic` decorated with
  `@polish_phase(name="path_thematic_annotation", depends_on=["atmospheric_annotation"], priority=3)`.
  One LLM call per path via `batch_llm_calls` produces `path_theme`
  + `path_mood`. R-5f.1: skips paths with <2 beats. R-5f.3: per-path
  failures log at WARNING and leave the fields unpopulated.

Registry-name note: spec calls them "Phase 5e" and "Phase 5f" sub-phases
of Phase 5 LLM Enrichment, but `_phase_5_llm_enrichment` already uses
internal sub-step labels 5a-5f (5e=feasibility, 5f=transitions). To
avoid that collision and keep them as discrete @polish_phase units,
the registry names are `atmospheric_annotation` and
`path_thematic_annotation`. Spec's "Phase 5e/5f" labels are conceptual;
rule numbers (R-5e.*, R-5f.*) refer to the spec rules these phases
implement.

**GROW** (`grow/llm_phases.py`):
- `_phase_4d_atmospheric` removed.
- `_phase_4e_path_arcs` removed.
- `transition_gaps.depends_on` updated from
  `["path_arcs", "entity_arcs"]` to `["entity_arcs"]`.
- Imports cleaned up (`build_narrative_frame`, `enrich_beat_line` no
  longer used in GROW).

**Prompts**:
- `prompts/templates/polish_phase5e_atmospheric.yaml` (verbatim body
  from grow_phase4d_atmospheric.yaml; only header renamed).
- `prompts/templates/polish_phase5f_path_thematic.yaml` (verbatim body
  from grow_phase4e_path_arcs.yaml; only header renamed).

**Tests**:
- `test_grow_stage.py::TestGrowStagePhaseOrder`: phase count 17 → 15;
  `atmospheric` and `path_arcs` removed from expected sequence.
- `test_grow_registry.py::TestGlobalRegistry`: same — 17 → 15.

Tests: 3566 unit tests pass; mypy + ruff clean.

Refs #1368.
