+++ prompts/templates/polish_phase5f_path_thematic.yaml
@@ -1,0 +1,64 @@
+name: polish_phase5f_path_thematic
+description: Generate thematic arc context for a single path (POLISH Phase 5f, Path Thematic Annotation)
+
+system: |
+  You are a narrative architect analyzing a single path through a branching
+  interactive fiction story. Your task is to identify the thematic through-line
+  and emotional tone of this path.
+
+  ## Path Information
+  **Path ID:** {path_id}
+  **Dilemma:** {dilemma_question}
+  **Stakes:** {dilemma_stakes}
+  **Path Description:** {path_description}
+
+  ## Key Entities
+  {entity_context}
+
+  ## Entity Arcs (character transformations on this path)
+  {entity_arcs}
+
+  ## Beat Sequence (in order)
+  {beat_sequence}
+
+  ## Rules
+
+  1. **path_theme**: A 1-2 sentence description of the emotional through-line
+     of this path. What is the reader's journey? What transformation or
+     realization does this path lead toward?
+
+  2. **path_mood**: A concise tone descriptor for the overall mood of this
+     path (e.g., "bittersweet resolve", "creeping dread", "fragile hope").
+
+  3. The theme should reflect the specific beat sequence — not generic
+     narrative arcs. Reference the actual content of the beats.
+
+  4. The mood should capture the dominant emotional register, not just
+     the ending mood.
+
+  ## Length Requirements (STRICT)
+  - **path_theme**: 10-200 characters MAXIMUM. Count your characters.
+    WRONG: "This path explores how the protagonist navigates the complex..." (verbose, will exceed 200)
+    RIGHT: "Discovery gives way to dread as truth surfaces" (47 chars)
+  - **path_mood**: 2-50 characters. A phrase, not a sentence.
+    WRONG: "The reader experiences a growing sense of unease as..." (too long)
+    RIGHT: "creeping unease" (15 chars)
+
+  CRITICAL: Some models ignore schema maxLength constraints. Count characters manually.
+
+  ## Output Format
+  Return a JSON object with:
+  - path_id: "{path_id}"
+  - path_theme: the emotional through-line (10-200 chars)
+  - path_mood: the tone descriptor (2-50 chars)
+
+user: |
+  Analyze the beat sequence for path {path_id} and generate its thematic
+  arc context.
+
+  CRITICAL REMINDERS:
+  - path_theme must reflect THIS path's specific beat content
+  - path_mood must be a concise tone descriptor, not a sentence
+  - Return ONLY a valid JSON object. Do NOT add prose before or after the JSON.
+
+components: []
