+++ tests/unit/test_grow_registry.py
@@ -216,11 +216,11 @@ class TestGrowPhaseDecorator:
 class TestGlobalRegistry:
     """Tests for the global registry populated by actual GROW phases."""
 
-    def test_global_registry_has_17_phases(self) -> None:
-        """All GROW phases are registered (17 after narrative_gaps moved to POLISH per #1368)."""
+    def test_global_registry_has_15_phases(self) -> None:
+        """GROW phases (15 after narrative_gaps/atmospheric/path_arcs moved to POLISH per #1368)."""
         registry = get_registry()
-        assert len(registry) == 17, (
-            f"Expected 17 phases, got {len(registry)}: {registry.phase_names}"
+        assert len(registry) == 15, (
+            f"Expected 15 phases, got {len(registry)}: {registry.phase_names}"
         )
 
     def test_global_registry_validates(self) -> None:
@@ -230,7 +230,7 @@ class TestGlobalRegistry:
         assert errors == [], f"Registry validation errors: {errors}"
 
     def test_global_registry_execution_order_matches_expected(self) -> None:
-        """Execution order matches the current phase structure (17 phases)."""
+        """Execution order matches the current phase structure (15 phases)."""
         expected = [
             "validate_dag",
             "intersections",
@@ -238,10 +238,10 @@ class TestGlobalRegistry:
             "resolve_temporal_hints",
             "interleave_beats",
             "scene_types",
-            # narrative_gaps moved to POLISH Phase 1a per PR #1366 / issue #1368
+            # narrative_gaps moved to POLISH Phase 1a (PR A of #1368)
             "pacing_gaps",
-            "atmospheric",
-            "path_arcs",
+            # atmospheric moved to POLISH Phase 5e (PR B of #1368)
+            # path_arcs moved to POLISH Phase 5f (PR B of #1368)
             "entity_arcs",
             "transition_gaps",
             "enumerate_arcs",
